+++ weblate/urls.py
@@ -176,6 +176,11 @@ urlpatterns = [
         name='edit_priority'
     ),
     url(
+        r'^source/(?P<pk>[0-9]+)/context/$',
+        weblate.trans.views.source.edit_context,
+        name='edit_context'
+    ),
+    url(
         r'^source/(?P<pk>[0-9]+)/check_flags/$',
         weblate.trans.views.source.edit_check_flags,
         name='edit_check_flags'